// src/shared/types.ts
export interface Employee {
  id?: string;
  employee_id: string; // Khóa chính
  name: string;
  email: string;
  password?: string | number;
  role: string; // "Manager", "Staff", "Admin"
  center_id: string;
  position: string;
  department: string;
  phone: string | number;
  annual_leave_balance: number;
  face_ref_url: string; // Avatar
  trusted_device_id: string;
  status: 'Active' | 'Inactive';
  direct_manager_id?: string;
  join_date?: string;
  theme_color?: string;
  fcm_tokens?: string[];
}
// ...
export interface LeaveRequest {
  id?: string; // Firestore Doc ID
  request_id: string;
  employee_id: string;
  name: string;
  created_at: string; // ISO String
  type: string;
  from_date: string; // YYYY-MM-DD
  to_date: string; // YYYY-MM-DD
  expiration_date?: string; // YYYY-MM-DD
  reason: string;
  status: 'Pending' | 'Approved' | 'Rejected';
  note?: string; // Ghi chú của quản lý
  manager_note?: string;
}

export interface Explanation {
  id?: string;
  employee_id: string;
  name: string;
  date: string; // The date being explained (YYYY-MM-DD)
  reason: string;
  status: 'Pending' | 'Approved' | 'Rejected';
  created_at: string;
  manager_note?: string;
}